+++ CHANGELOG.md
@@ -4,7 +4,19 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-## [0.0.2] - 2018-03-12
+## [0.1.1] - 2018-03-17
+
+### Add Result
+
+- Add result return in cxapiSelectNegativeFaceCmd.
+- Maya C++ API #8
+
+### Add Flag
+
+- Add a flag to cxapiSelectNegativeFaceCmd - tolerance.  
+- Maya C++ API #9
+
+## [0.1.0] - 2018-03-12
 
 ### Added